fix: Moved query client to index from App.
