+++ src/index.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { QueryClient, QueryClientProvider } from 'react-query';
 import './index.css';
 import 'react-loading-skeleton/dist/skeleton.css';
 
@@ -18,31 +19,44 @@ import PoolOverview from 'pages/pool/components/PoolsOverview/Pools/PoolOverview
 import ManagePosition from 'pages/pool/components/ManagePosition';
 import IncreaseLiquidity from 'pages/pool/components/IncreaseLiquidity';
 
+// Refetch the data after 1 hour.
+const oneHourInMs = 60 * 60 * 1000;
+const queryClientOptions = {
+  defaultOptions: {
+    queries: {
+      staleTime: oneHourInMs,
+    },
+  },
+};
+const queryClient = new QueryClient(queryClientOptions);
+
 ReactDOM.render(
   <React.StrictMode>
     <ToastContainer className="font-sans font-semibold" />
-    <AppProviders>
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<App />}>
-            <Route path="/" element={<Navigate replace to="/bridge" />} />
-            <Route path="/bridge" element={<Bridge />} />
-            <Route path="/pool" element={<Pool />}>
-              <Route path="" element={<PoolsOverview />} />
-              <Route path="add-liquidity" element={<AddLiquidity />} />
-              <Route
-                path="manage-position/:chainId/:positionId"
-                element={<ManagePosition />}
-              />
-              <Route
-                path="increase-liquidity/:chainId/:positionId"
-                element={<IncreaseLiquidity />}
-              />
+    <QueryClientProvider client={queryClient}>
+      <AppProviders>
+        <BrowserRouter>
+          <Routes>
+            <Route path="/" element={<App />}>
+              <Route path="/" element={<Navigate replace to="/bridge" />} />
+              <Route path="/bridge" element={<Bridge />} />
+              <Route path="/pool" element={<Pool />}>
+                <Route path="" element={<PoolsOverview />} />
+                <Route path="add-liquidity" element={<AddLiquidity />} />
+                <Route
+                  path="manage-position/:chainId/:positionId"
+                  element={<ManagePosition />}
+                />
+                <Route
+                  path="increase-liquidity/:chainId/:positionId"
+                  element={<IncreaseLiquidity />}
+                />
+              </Route>
             </Route>
-          </Route>
-        </Routes>
-      </BrowserRouter>
-    </AppProviders>
+          </Routes>
+        </BrowserRouter>
+      </AppProviders>
+    </QueryClientProvider>
   </React.StrictMode>,
   document.getElementById('root'),
 );
